if the windows update service is disabled then assume that windows updates is turned off.
to test this, we tried to find a way to update windows while the update service is turned off.
we were unsuccesful. Therefore, we concluded that the checking if the windows update service
is turned off, then there is no way for windows to update, and the additional test fixtures
(that were failing in gcp and passing elsewhere) are not neccessary.
[#151438181]

Signed-off-by: Paul Nikonowicz <pnikonowicz@pivotal.io>