Ensure that Write-Log writes console messages through Write-Host
Write-Host was removed because it was believed that just outputing the
message withtout it would always write it to console. However, in
certain scenarios, when Write-Log was used in a function that returned a
value, the logged message was returned as the result of the function.
As such, the code evaluating the result of such function would
see the message as part of the result and cause unexpected behaviors
(such as incorrectly validating Powershell version on the VM).

[#161899385]