Relocates resources from incubator org
We discovered that as of a few days ago, the `cloudfoundry-incubator` ->
`cloudfoundry` redirect for several Bosh Windows Stemcell Creator
resources has stopped working. No idea why, but we're pretty sure that
it prevents Packer from downloading the scripts needed to correctly set
up WinRM on the Windows VM that it creates to build the stemcell.

So, this commit corrects the relevant URLs, which will _hopefully_ make
the Windows Stemcell builder pipeline work again.

Signed-off-by: Rajan Agaskar <ragaskar@vmware.com>