Fix test-stemcell-builder
The implementation of the AWS AMI aggregator uses Ruby's Dir.entries to
determine the final file name of the aggregated stemcell. After updating
the Ruby version used by the Docker image, this ordering appears to have
changed, leading to test failures.

Based on looking at how the final file name is treated in the rest of
the pipeline, we do not believe that the file selected to generate the
final file name matters. Thus, we do not think the test around the exact
file name is significant.

Signed-off-by: Brian Upton <bupton@vmware.com>