Shorten capture_name_prefix
While including "stemcell" and the OS and version makes it obvious what
the final VHD file is, there is a 23 character limit for the
capture_name_prefix in Packer. The storage account +
capture_container_name should already make it obvious that it is a
stemcell, and is of a particular OS, so only include the version as the
prefix.