Remove usage of `bosh-agent` submodule
Previously the code to assemble the zip file of agent assets relied on
files from the `bosh-agent-index` as well as the git version info to
checkout the `bosh-agent` submodule of this repo to the correct version,
and then built `pipe.exe` and copied `service_wrapper.xml`.

This commit changes the CI task's `run.sh` to instead neither rename
`bosh-agent.exe`, and `pipe.exe`, nor to update the submodule commit,
and instead updates the rake task to fetch these assets directly from
the `bosh-agent-index` where the are now being added[1], and the git
submodule at srrc/github.com/cloudfoundry/bosh-agent/ has been removed.

[1] example: https://github.com/cloudfoundry/bosh-agent-index/blob/master/v2.713.0.meta4