Stop populating `apply_spec.yml` - AWS, Azure, GCP
A file called `apply_spec.yml` had historically been placed onto the
Windows stemcell on AWS, Azure, GCP, and possibly prior to `stembuild`
on vSphere as well.

Given that the vSphere Windows stemcells have shipped without this file
for many years it does not appear to be valuable.

For consistency this is being removed everywhere instead of propigated
to vSphere.