Powershel: refactor Enable-LocalSecurityPolicy
This method was previously not tested, only it's invocation. This commit
changes the contents of Enable-LocalSecurityPolicy to move several
external checks (os, presence of LGPO.exe) inside the method and tests
its invocation of LGPO.exe explicitly. This change switches from "bare"
invocations of LGPO.exe to the use of `Invoke-Expression` which should
be equivalent for the commands in question.