Stembuild: upload all powershell modules in setup
The Construct test suite has been relying on the BOSH.WinRM.psm1 module
and not uploading the BOSH.Utils.psm1 module. When the WinRM code was
updated to remove a duplicated `Write-Log` method this caused different
behavior when it was executed without access to BOSH.Utils.psm1.

This commit updates the test setup to upload both BOSH.WinRM.psm1 and
BOSH.Utils.psm1.