acceptance_test: remove auto-generated pre-start content
It seems like there was at one point some code in
acceptance_test/main_test.go which updated the simple-job/pre-start.ps1
to include a counter which was incremented when redeploys happened.
While some related code exists (file Open) nothing is ever written to
that file. This change removes the seemingly auto-generated content in
pre-start.ps1 and replaces it with something that prints a date.