Move GCP publishing to the end of the pipeline
Previously, we were immediately creating a public image for GCP when we
built the stemcell. This is not allowed in labs, as we need to publish
to a separate account. This commit moves the publishing step to the end
of the pipeline, and replaces the previous publishing ruby code w/ a
bash script.

We have purposefully removed labeling because the dev and public images
now live in different accounts.

Signed-off-by: Brian Upton <brian.upton@broadcom.com>