ci: install-windows-updates fix `get_windows_updates_remaining` loop
It appears that the new(er) version of `govc` does not retain/return the
exit code of a process when checking `guest.ps`, so that if a process
that was `guest.start`'ed exits before the process can be checked the
resulting JSON response will not contain an exit code (or at least that
appears to be what is happening).

In stead of breaking out of the loop to check for remaining updates when
there is _any_ response, we now check to see if the response is null or
not.